hopeful fix for memory leak:delete preOR objects at the end of NewObjectDef

+++ WorkDir/MyAnalysis/util/testRunTest.cxx
@@ -195,75 +195,75 @@ std::string get_file_type(const std::string& s)
 
     std::string fileType = "UNKNOWN";
 
-    int found_TRUTH1 = s.find("TRUTH1");
+    size_t found_TRUTH1 = s.find("TRUTH1");
     if (found_TRUTH1 != std::string::npos){
       fileType = "TRUTH1";
     }
-    int found_TRUTH3 = s.find("TRUTH3");
+    size_t found_TRUTH3 = s.find("TRUTH3");
     if (found_TRUTH3 != std::string::npos){
       fileType = "TRUTH3";
     }
-    int found_data15 = s.find("data15");
+    size_t found_data15 = s.find("data15");
     if (found_data15 != std::string::npos){
       fileType = "DATA15";
-        int found_SUSY5 = s.find("SUSY5");
+        size_t found_SUSY5 = s.find("SUSY5");
         if (found_SUSY5 != std::string::npos){
         fileType += ".SUSY5";
         }
-        int found_SUSY7 = s.find("SUSY7");
+        size_t found_SUSY7 = s.find("SUSY7");
         if (found_SUSY7 != std::string::npos){
         fileType += ".SUSY7";
         }
     }
     
-    int found_data16 = s.find("data16");
+    size_t found_data16 = s.find("data16");
     if (found_data16 != std::string::npos){
       fileType = "DATA16";
-        int found_SUSY5 = s.find("SUSY5");
+        size_t found_SUSY5 = s.find("SUSY5");
         if (found_SUSY5 != std::string::npos){
         fileType += ".SUSY5";
         }
-        int found_SUSY7 = s.find("SUSY7");
+        size_t found_SUSY7 = s.find("SUSY7");
         if (found_SUSY7 != std::string::npos){
         fileType += ".SUSY7";
         }
     }
     
-    int found_data17 = s.find("data17");
+    size_t found_data17 = s.find("data17");
     if (found_data17 != std::string::npos){
       fileType = "DATA17";
-        int found_SUSY5 = s.find("SUSY5");
+        size_t found_SUSY5 = s.find("SUSY5");
         if (found_SUSY5 != std::string::npos){
         fileType += ".SUSY5";
         }
-        int found_SUSY7 = s.find("SUSY7");
+        size_t found_SUSY7 = s.find("SUSY7");
         if (found_SUSY7 != std::string::npos){
         fileType += ".SUSY7";
         }
     }
     
-    int found_data18 = s.find("data18");
+    size_t found_data18 = s.find("data18");
     if (found_data18 != std::string::npos){
       fileType = "DATA18";
-        int found_SUSY5 = s.find("SUSY5");
+        size_t found_SUSY5 = s.find("SUSY5");
         if (found_SUSY5 != std::string::npos){
         fileType += ".SUSY5";
         }
-        int found_SUSY7 = s.find("SUSY7");
+        size_t found_SUSY7 = s.find("SUSY7");
         if (found_SUSY7 != std::string::npos){
         fileType += ".SUSY7";
         }
     }
     
-    int found_mc16a = s.find("r9364");
+    size_t found_mc16a = s.find("r9364");
     if(found_mc16a != std::string::npos){
       fileType = "MC16a";
     }
-    int found_mc16d = s.find("r10201");
+    size_t found_mc16d = s.find("r10201");
     if(found_mc16d != std::string::npos){
       fileType = "MC16d";
     }
-    int found_mc16e = s.find("r10724");
+    size_t found_mc16e = s.find("r10724");
     if(found_mc16e != std::string::npos){
       fileType = "MC16e";
     }
@@ -274,7 +274,7 @@ std::string get_file_type(const std::string& s)
 bool is_multiple_submission(const std::string& s) 
 {
 
-    int multi_sub = s.find(",");
+    size_t multi_sub = s.find(",");
     if (multi_sub != std::string::npos){
         return true;
     }
